netstacklat: Refactor eBPF latency recording to be more flexible
Refactor the eBPF code for recording latency to be more general, to
make it reuseable across a wider range of context. Specifically,
change the record_current_netstacklat(skb, hook) function to
record_latency_since(tstamp, hook), so that one does not have to
supply an skb.

This allows the core logic for recording the latency be reused even
for hooks that may not directly have access to the skb, but still have
a relevant timestamp to compare against.

Furthermore, this allows reducing the overhead for some hooks, as each
hook can provide access to the skb->tstamp in an optimized
manner. Specifically, fentry hooks can directly access the memory of
the provided arguments, so they do not need to use the SKB_CORE_READ()
macro. Fentry hooks can thereby avoid the overhead from the
SKB_CORE_READ probing the memory, which is only needed for hooks that
do not have direct memory access for its arguments (like
e.g. raw_tracepoints).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>