netstacklat: Add/change probe points
Add several new probes at different parts of the networks stack.
Remove one previous probe (tcp_v4_do_rcv) which did not fit in with
the new selection of hooks.

The following hooks are now present:

- ip-start: Packet has reached the IP-stack (past tc-layer).
  Hooks fentry:ip_rcv_core, ip6_rcv_core
- conntrack: Packet has passed through conntrack.
  Hooks fexit:ipv4_conntrack_in, ipv6_conntrack_in
- iptables: Packet has passed through iptables
  Hooks fexit:ipt_do_table, ip6t_do_table
- tcp-socket-queued: Packet has been queued to TCP socket
  Hooks fexit:tcp_data_queue
- udp-socket-queued: Packed has been queued to UDP socket
  Hooks fexit: udp_queue_one_skb, udpv6_queue_rcv_one_skb
- tcp-socket-read: Packet data read from TCP socket
  Hooks fentry:tcp_recv_timestamp
- udp-socket-read: Packet data read from UDP socket
  Hooks fentry:skb_consume_udp

Furthermore, employ a new naming strategy, presenting the user with
intended layer/component that is probed (the names before the colon in
the list above) rather than the function name(s). This is intended to
make it easier for the user to understand the purpose of the probe,
and also allows several underlying hooks (to e.g. handle IPv4 and IPv6
paths) for each probe point.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>